cloud_function/requirements.txt
functions-framework==3.5.0
google-cloud-bigquery==3.14.1